+++ projects/ng-porcelain/src/lib/search-input/search-input.module.spec.ts
@@ -1,0 +1,13 @@
+import { SearchInputModule } from './search-input.module';
+
+describe('SearchInputModule', () => {
+	let searchInputModule: SearchInputModule;
+
+	beforeEach(() => {
+		searchInputModule = new SearchInputModule();
+	});
+
+	it('should create an instance', () => {
+		expect(searchInputModule).toBeTruthy();
+	});
+});