+++ projects/ng-porcelain/src/public_api.ts
@@ -16,6 +16,9 @@ export { RefinersComponent } from './lib/refiners/refiners/refiners.component';
 export { SimpleRefinerModule } from './lib/simple-refiner/simple-refiner.module';
 export { SimpleRefinerComponent } from './lib/simple-refiner/simple-refiner/simple-refiner.component';
 
+export { SearchInputModule } from './lib/search-input/search-input.module';
+export { SearchInputComponent } from './lib/search-input/search-input/search-input.component';
+
 export { SpinnerModule } from './lib/spinner/spinner.module';
 export { SpinnerComponent } from './lib/spinner/spinner/spinner.component';